renamed boardBuilder.ts to boardUtils.ts FreeCodeCamp - Data Visualization Certification - React Projects
